# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAFE 43 pastry shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Category = (typeof Category)[keyof typeof Category];

export const Category = {
  cakes: "cakes",
  pastries: "pastries",
  drinks: "drinks",
  chocolate: "chocolate",
} as const;

export interface MenuItem {
  id: number;
  name: string;
  nameKa: string;
  description: string;
  descriptionKa: string;
  price: number;
  category: Category;
  imageUrl: string;
  featured: boolean;
  createdAt: string;
}

export interface CreateMenuItemBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  nameKa: string;
  description: string;
  descriptionKa: string;
  /** @minimum 0 */
  price: number;
  category: Category;
  imageUrl: string;
  featured?: boolean;
}

export interface UpdateMenuItemBody {
  name?: string;
  nameKa?: string;
  description?: string;
  descriptionKa?: string;
  price?: number;
  category?: Category;
  imageUrl?: string;
  featured?: boolean;
}

export interface CategorySummary {
  category: Category;
  count: number;
  sampleImageUrl: string;
}

export interface Review {
  id: number;
  username: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  text: string;
  createdAt: string;
}

export interface CreateReviewBody {
  /**
   * @minLength 1
   * @maxLength 60
   */
  username: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /**
   * @minLength 1
   * @maxLength 1000
   */
  text: string;
}

export interface ReviewSummary {
  averageRating: number;
  count: number;
}

export interface GalleryPhoto {
  id: number;
  imageUrl: string;
  caption: string;
}

export interface OrderItemInput {
  menuItemId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderItem {
  menuItemId: number;
  name: string;
  unitPrice: number;
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  ready: "ready",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  address: string;
  notes: string;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  createdAt: string;
}

export interface CreateOrderBody {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 4 */
  phone: string;
  /** @minLength 1 */
  address: string;
  notes?: string;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export interface ShopStatus {
  isOpen: boolean;
  opensAt: string;
  closesAt: string;
  currentLocalTime: string;
}

export interface ShopInfo {
  name: string;
  tagline: string;
  address: string;
  city: string;
  phone: string;
  email: string;
  rating: number;
  reviewCount: number;
  latitude: number;
  longitude: number;
  hours: string;
}

export type ListMenuParams = {
  category?: Category;
  search?: string;
};