# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CAFE 43 pastry shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu:
    get:
      operationId: listMenu
      tags: [menu]
      summary: List menu items
      parameters:
        - in: query
          name: category
          required: false
          schema:
            $ref: "#/components/schemas/Category"
        - in: query
          name: search
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create menu item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single menu item
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update menu item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete menu item (admin)
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Category:
      type: string
      enum: [cakes, pastries, drinks, chocolate]
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameKa:
          type: string
        description:
          type: string
        descriptionKa:
          type: string
        price:
          type: number
        category:
          $ref: "#/components/schemas/Category"
        imageUrl:
          type: string
        featured:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, nameKa, description, descriptionKa, price, category, imageUrl, featured, createdAt]
    CreateMenuItemBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        nameKa:
          type: string
          minLength: 1
        description:
          type: string
        descriptionKa:
          type: string
        price:
          type: number
          minimum: 0
        category:
          $ref: "#/components/schemas/Category"
        imageUrl:
          type: string
        featured:
          type: boolean
      required: [name, nameKa, description, descriptionKa, price, category, imageUrl]
    UpdateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        nameKa:
          type: string
        description:
          type: string
        descriptionKa:
          type: string
        price:
          type: number
        category:
          $ref: "#/components/schemas/Category"
        imageUrl:
          type: string
        featured:
          type: boolean
...