# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAFE 43 pastry shop API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List menu items
 */
export const ListMenuQueryParams = zod.object({
  category: zod.enum(["cakes", "pastries", "drinks", "chocolate"]).optional(),
  search: zod.coerce.string().optional(),
});

export const ListMenuResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameKa: zod.string(),
  description: zod.string(),
  descriptionKa: zod.string(),
  price: zod.number(),
  category: zod.enum(["cakes", "pastries", "drinks", "chocolate"]),
  imageUrl: zod.string(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListMenuResponse = zod.array(ListMenuResponseItem);

/**
 * @summary Create menu item (admin)
 */

export const createMenuItemBodyPriceMin = 0;

export const CreateMenuItemBody = zod.object({
  name: zod.string().min(1),
  nameKa: zod.string().min(1),
  description: zod.string(),
  descriptionKa: zod.string(),
  price: zod.number().min(createMenuItemBodyPriceMin),
  category: zod.enum(["cakes", "pastries", "drinks", "chocolate"]),
  imageUrl: zod.string(),
  featured: zod.boolean().optional(),
});

/**
 * @summary Get a single menu item
 */
export const GetMenuItemParams = zod.object({
  id: zod.coerce.number(),
});

export const GetMenuItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameKa: zod.string(),
  description: zod.string(),
  descriptionKa: zod.string(),
  price: zod.number(),
  category: zod.enum(["cakes", "pastries", "drinks", "chocolate"]),
  imageUrl: zod.string(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update menu item (admin)
 */
export const UpdateMenuItemParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateMenuItemBody = zod.object({
  name: zod.string().optional(),
  nameKa: zod.string().optional(),
  description: zod.string().optional(),
  descriptionKa: zod.string().optional(),
  price: zod.number().optional(),
  category: zod.enum(["cakes", "pastries", "drinks", "chocolate"]).optional(),
  imageUrl: zod.string().optional(),
  featured: zod.boolean().optional(),
});

export const UpdateMenuItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameKa: zod.string(),
  description: zod.string(),
  descriptionKa: zod.string(),
  price: zod.number(),
  category: zod.enum(["cakes", "pastries", "drinks", "chocolate"]),
  imageUrl: zod.string(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete menu item (admin)
 */
export const DeleteMenuItemParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Featured / top items for the home page
 */
export const ListFeaturedMenuResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameKa: zod.string(),
  description: zod.string(),
  descriptionKa: zod.string(),
  price: zod.number(),
  category: zod.enum(["cakes", "pastries", "drinks", "chocolate"]),
  imageUrl: zod.string(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListFeaturedMenuResponse = zod.array(ListFeaturedMenuResponseItem);

/**
 * @summary Category counts and example item per category
 */
export const ListCategoriesSummaryResponseItem = zod.object({
  category: zod.enum(["cakes", "pastries", "drinks", "chocolate"]),
  count: zod.number(),
  sampleImageUrl: zod.string(),
});
export const ListCategoriesSummaryResponse = zod.array(
  ListCategoriesSummaryResponseItem,
);

/**
 * @summary List reviews
 */
export const listReviewsResponseRatingMax = 5;

export const ListReviewsResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  rating: zod.number().min(1).max(listReviewsResponseRatingMax),
  text: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListReviewsResponse = zod.array(ListReviewsResponseItem);

/**
 * @summary Submit a review
 */
export const createReviewBodyUsernameMax = 60;

export const createReviewBodyRatingMax = 5;

export const createReviewBodyTextMax = 1000;

export const CreateReviewBody = zod.object({
  username: zod.string().min(1).max(createReviewBodyUsernameMax),
  rating: zod.number().min(1).max(createReviewBodyRatingMax),
  text: zod.string().min(1).max(createReviewBodyTextMax),
});

/**
 * @summary Average rating and review count
 */
export const GetReviewSummaryResponse = zod.object({
  averageRating: zod.number(),
  count: zod.number(),
});

/**
 * @summary List gallery images
 */
export const ListGalleryResponseItem = zod.object({
  id: zod.number(),
  imageUrl: zod.string(),
  caption: zod.string(),
});
export const ListGalleryResponse = zod.array(ListGalleryResponseItem);

/**
 * @summary List recent orders (admin)
 */
export const ListOrdersResponseItem = zod.object({
  id: zod.number(),
  customerName: zod.string(),
  phone: zod.string(),
  address: zod.string(),
  notes: zod.string(),
  items: zod.array(
    zod.object({
      menuItemId: zod.number(),
      name: zod.string(),
      unitPrice: zod.number(),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
  status: zod.enum(["pending", "confirmed", "ready", "delivered", "cancelled"]),
  createdAt: zod.coerce.date(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Place an order
 */

export const createOrderBodyPhoneMin = 4;

export const CreateOrderBody = zod.object({
  customerName: zod.string().min(1),
  phone: zod.string().min(createOrderBodyPhoneMin),
  address: zod.string().min(1),
  notes: zod.string().optional(),
  items: zod
    .array(
      zod.object({
        menuItemId: zod.number(),
        quantity: zod.number().min(1),
      }),
    )
    .min(1),
});

/**
 * @summary Open/closed status with next opening time
 */
export const GetShopStatusResponse = zod.object({
  isOpen: zod.boolean(),
  opensAt: zod.string(),
  closesAt: zod.string(),
  currentLocalTime: zod.string(),
});

/**
 * @summary Shop name, address, contact, hours
 */
export const GetShopInfoResponse = zod.object({
  name: zod.string(),
  tagline: zod.string(),
  address: zod.string(),
  city: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  rating: zod.number(),
  reviewCount: zod.number(),
  latitude: zod.number(),
  longitude: zod.number(),
  hours: zod.string(),
});