# lib/api-zod/src/generated/types/createMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAFE 43 pastry shop API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface CreateMenuItemBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  nameKa: string;
  description: string;
  descriptionKa: string;
  /** @minimum 0 */
  price: number;
  category: Category;
  imageUrl: string;
  featured?: boolean;
}